Move StatusDisplay to common lib
