+++ lib/StatusDisplay/StatusDisplay.cpp
